# animals-place-api/animals-place/openapi.yml
openapi: 3.1.0
info:
  title: Animals Place API
  description: API for an Animal Place Meeting app develloped in the ArchiOWeb course
  termsOfService: http://swagger.io/terms/
  contact:
    email: jorane.membrez@heig-vd.ch
  version: 2.0.0
  externalDocs:
    description: Link to the Github repository
    url: https://github.com/JoraneMembrez/archiOWeb-animalsPlace
servers:
  - url: http://localhost:8000
    description: Local server
  - url: https://archioweb-animalsplace.onrender.com/
    description: Heroku server
...
paths:
...
  /meetings/like/{animalID} :
    post :
      tags:
       - meetings
      summary: Like an animal by its ID
      description: You must be authenticate to execute this query
      parameters:
        - in: path
          name: animalID
          schema:
            type: string
            required: true
          description: The ID of the animal to like
      requestBody :
        required : true
        content:
          application/json:
            schema:
              type: object
              properties:
                animalUserID:
                  type: string
                  example: 6375ef6d1d458a3858c3453e
      responses:
        "200":
          description: Successful response
          content:
            application/json:
             schema:
               type: object
               properties :
                 animal :
                  $ref: "#/components/schemas/Animal"
        "401":
          $ref: "#/components/responses/401Unauthorized"
        "404":
          $ref: "#/components/responses/404NotFound"
        "400" :
          $ref: "#/components/responses/400BadRequest"
      security:
        - bearerAuth: []
  /meetings :
    get :
      tags:
       - meetings
      summary: Get all meetings that belong to the current user
      description: You must be authenticate to execute this query or be an admin
      responses: 
        "200":
          description: Successful response
          content:
            application/json:
             schema:
               type: object
               properties :
                 meetings :
                  type: array
                  items :
                    $ref: "#/components/schemas/Meeting"
               currentPage :
                 type: integer
               totalPages:
                  type: integer
        "401":
          $ref: "#/components/responses/401Unauthorized"
      security:
        - bearerAuth: []
  meetings/count :
    get :
      tags:
       - meetings
      summary: Get the number of meetings that one user has
      description: You must be authenticate to execute this query
      responses: 
        "200":
          description: Successful response
          content:
            application/json:
             schema:
               type: object
               properties :
                 count :
                  type: integer
        "401":
          $ref: "#/components/responses/401Unauthorized"
      security:
        - bearerAuth: []
  /meetings/{meetingID} :
    delete : 
        tags:
          - meetings
        summary: Delete a meeting by its ID
        description: You must be authenticate to execute this query or be an admin
        parameters:
          - in: path
            name: meetingID
            schema:
              type: string
              required: true
              example: 6375ef6d1d458a3858c3453e
            description: The ID of the meeting to delete
        responses:
          "200":
            description: Meeting deleted
            content:
              application/json:
                example: { "message": "Meeting deleted" }

          "403":   
           description : You are not authorized to access this resource
           content:
              application/json:
                example: { "message": "Forbidden" }
          
          "404":
            description : Not found
            content:
              application/json:
                example: { "message": "Meeting not found" }
        security:
          - bearerAuth: []
components:
  responses:
    401Unauthorized:
      description: L'en-tête d'autorisation est manquante
...
    404NotFound:
      description: La ressource recherchée n'existe pas
...
    400BadRequest:
      description: La demande n'est pas valide
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT # optional, for documentation purposes only
  schemas :
    UserBasic:
      type: object
      properties:
        _id:
          type: string
          example: 6375ef6d1d458a3858c3453e
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        email:
          type: string
          example: john.doe@gmail.com
        password:
          type: string
          example: "1234"
        registrationDate:
          type: string
          format: date-time
          example: 2021-04-13T14:30:00.000Z
        address :
          type: string
          example: "Rue de la gare 12, 2800 Delémont"
        animals :
          type: Object
          $ref : "#/components/schemas/Animal"
        role:
          type: string
          enum:
           - user
           - admin
          default: user
        location:
          type:
            type: string
            required: true
            enum:
              - Point
            default: Point
          coordinates:
            type:
             - number
            required: true
            validate:
              validator: validateGeoJsonCoordinates
              message: "{VALUE} is not a valid longitude/latitude(/altitude) coordinates array"
          default:
            - 6.647778558579233
            - 46.78060279685718 
...
    Animal :
      allOf :
        - type : object
          properties :
            _id:
              type: string
              example: 6375ef6d1d458a3858c3453e
            name : 
              type : string
              example : "Bobby"
            species :
              enum : 
                - chien
                - chat
                - lapin
                - furet
                - hamster
                - oiseau
                - tortue
                - poisson
                - souris
                - camélémon
                - serpent
                - araigné
                - autre
              example : chien
            age :
              type : integer
              example: 5
            gender :
              enum :
                - Male
                - Female
                - Other
            profilePictureUrl :
              type : string
            description :
                type : string
                example : "Bobby like painting and eating"
            picturesUrl :
                type : array
                items :
                  type : string
                example : [https://www.google.com/url?sa=i&url=https%3A%2F%2Fwww.woopets.fr%2Fchien%2F]
            location :
                type : String,
                example : "Rue de la gare 12, 2800 Delémont"
            favoriteActivites :
                type : array
                items :
                  type : string
                example : ["painting", "eating"]
            matches :
                type : array
                items :
                  $ref : "#/components/schemas/Animal"
                example : ["6375ef6d1d458a3858c3453e"]
            createdDate :
                type : string
                format : date-time
                example : 2021-04-13T14:30:00.000Z
            owner :
                $ref : "#/components/schemas/UserBasic"
                example : "6375ef6d1d458a3858c3453e"
            animal_liked :
                type : array
                items :
                  $ref : "#/components/schemas/Animal"
                example : ["6375ef6d1d458a3858c3453e"]
    Meeting :
      allOf :
        - type : object
          properties :
            _id:
              type: string
              example: 6375ef6d1d458a3858c3453e
            owner :
              $ref : "#/components/schemas/UserBasic"
              example : "6375ef6d1d458a3858c3453e"
            animal1 :
              $ref : "#/components/schemas/Animal"
              example : "6375ef6d1d458a3858c3453e"
            animal2 :
              $ref : "#/components/schemas/Animal"
              example : "6375ef6d1d458a3858c3457e"
            date :
              type : string
              format : date-time
              example : 2021-04-13T14:30:00.000Z
            location :
              type : string
              example : "Rue de la gare 12, 2800 Delémont"
            description :
              type : string
              example : "La rencontre aura lieu à la gare de Delémont"
            createdDate :
              type : string
              format : date-time
              example : 2021-04-13T14:30:00.000Z
...